# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lintel",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/approval-requests": {
      "post": {
        "tags": [
          "approval-requests"
        ],
        "summary": "Create Approval Request",
        "operationId": "approval-requests_create_approval_request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApprovalRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approval-Requests Create Approval Request"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "approval-requests"
        ],
        "summary": "List Approval Requests",
        "operationId": "approval-requests_list_approval_requests",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApprovalStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Approval-Requests List Approval Requests"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/approval-requests/{approval_id}": {
      "get": {
        "tags": [
          "approval-requests"
        ],
        "summary": "Get Approval Request",
        "operationId": "approval-requests_get_approval_request",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approval-Requests Get Approval Request"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approval-requests/{approval_id}/approve": {
      "post": {
        "tags": [
          "approval-requests"
        ],
        "summary": "Approve Approval Request",
        "operationId": "approval-requests_approve_approval_request",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approval-Requests Approve Approval Request"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approval-requests/{approval_id}/reject": {
      "post": {
        "tags": [
          "approval-requests"
        ],
        "summary": "Reject Approval Request",
        "operationId": "approval-requests_reject_approval_request",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approval-Requests Reject Approval Request"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalStatus": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "rejected",
          "expired"
        ],
        "title": "ApprovalStatus"
      },
...
      "CreateApprovalRequestBody": {
        "properties": {
          "approval_id": {
            "type": "string",
            "title": "Approval Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "gate_type": {
            "type": "string",
            "title": "Gate Type"
          },
          "requested_by": {
            "type": "string",
            "title": "Requested By",
            "default": ""
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "gate_type"
        ],
        "title": "CreateApprovalRequestBody"
      },
...
      "DecisionBody": {
        "properties": {
          "decided_by": {
            "type": "string",
            "title": "Decided By"
          }
        },
        "type": "object",
        "required": [
          "decided_by"
        ],
        "title": "DecisionBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RejectBody": {
        "properties": {
          "decided_by": {
            "type": "string",
            "title": "Decided By"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "decided_by"
        ],
        "title": "RejectBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}